@@ -6,6 +6,11 @@
 	import Header from './Header.svelte'
 
 	import '../app.css'
+	interface Props {
+		children?: import('svelte').Snippet;
+	}
+
+	let { children }: Props = $props();
 </script>
 
 {#if $isLoading || $langPreference === null}
@@ -15,7 +20,7 @@
 		<Header />
 
 		<main>
-			<slot />
+			{@render children?.()}
 		</main>
 
 		<Footer />
